requirements.txt
langchain-core
langchain-openai
langchain-community
langchain-chroma
langchain-huggingface
langgraph
playwright
pytz
requests
beautifulsoup4
lxml
psutil
sentence-transformers
chromadb
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
